pyproject.toml
[project]
name = "patent_crew"
version = "0.1.0"
description = "patent-crew using crewAI"
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.121.1",
    "arize-phoenix >= 7.0.1",
    "docling>=2.34.0",
    "python-dotenv>=1.1.0",
    "crewai-tools>=0.45.0",
    "agentops>=0.4.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
